+++ README.md
@@ -60,17 +60,17 @@ const balance = uint256('balance');
 const signedValue = int256('signed_value');
 ```
 
-## Middleware
+## Middlewares
 
 The library provides a middleware for authentication.
 
 It is based on the `X-IDX-AUTHENTICATED-API-KEY-NAME` header that is automatically set when valid `api_key` query parameter or `Authorization` header is present.
 
 ```typescript
-import { create, middleware } from '@duneanalytics/sim-idx';
+import { create, middlewares } from '@duneanalytics/sim-idx';
 
 const app = create();
-app.use('*', middleware.authentication);
+app.use('*', middlewares.authentication);
 ```
 
 ## Environment Variables
