+++ src/index.ts
@@ -1,6 +1,6 @@
 import * as db from './db';
 import * as App from './app';
 import * as types from './types';
-import * as middleware from './middleware';
+import * as middleware from './middlewares';
 
 export { App, db, types, middleware };
